arrumando free bullets inimigos

+++ Bullet.h
@@ -5,13 +5,13 @@
 #define BULLET_COOLDOWN 0
 
 typedef struct {																																//Definição da estrutura de um pŕojétil (!)
-	unsigned short x;																															//Local, no eixo x, onde se localiza a bala (!)
-	unsigned short y;																															//Local, no eixo y, onde se localiza a bala (!)
+	int x;																															//Local, no eixo x, onde se localiza a bala (!)
+	int y;																															//Local, no eixo y, onde se localiza a bala (!)
 	unsigned char trajectory;				
 	struct bullet *next; 																														//Próxima bala, se houver, formando uma lista encadeada (!)
 } bullet;																																		//Definição do nome da estrutura (!)
 
-bullet* bullet_create(unsigned short x, unsigned short y, char trajectory, bullet *next);												//Protótipo da função de criação de uma bala (!)
+bullet* bullet_create(int x, int y, char trajectory, bullet *next);												//Protótipo da função de criação de uma bala (!)
 void bullet_move(bullet *elements);			
 int has_shot_column(bullet* elements, int column);																									//Protótipo da função de movimentação de uma bala (!)	
 void bullet_destroy(bullet* element);																											//Protótipo da função de destruição de uma bala (!)
